+++ src-tauri/src/domain/interfaces.rs
@@ -5,7 +5,7 @@ use crate::domain::models::InputMethod;
 pub trait InputSource: Send {
     /// Returns a bitmask of pressed buttons.
     /// Returns Err(InputError::Disconnected) if device is disconnected.
-    fn get_state(&mut self, controller_index: u32) -> Result<u16, InputError>;
+    fn get_state(&mut self, controller_index: u32) -> Result<u32, InputError>;
 
     /// Optional: updates the input method if the source supports switching.
     fn set_input_method(&mut self, _method: InputMethod) {}
